Fix a bug which stopped PiP in certain cases
This commit changes some configruation so that the Chrome extension applies to more pages on the Channel 4 domain. There was a particular scenario where the extension wasn't executing when it should have been.

When navigating from the Channel 4 homepage, to a programme page, and then on to watch an episode, the extension wasn't firing. This was due to the navigation between Programme page => Playback page happening dynamically without a full page reload. This meant the extension didn't execute and re-enable PiP on the `<video>` element on the page.